Fix for older Ruby versions where Regexp#union does not handle an Array type.
This is to address problem where older versions of Ruby had Regexp#union
implementation that only accepts a list of bare strings instead of an array.

In order to fix, we simply splay an array so it will appear as a list of
bare strings and therefore Regexp#union will work.

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>