Add facts for getting UIDs and GIDs of Linux users
As some software requires specifying UID and GID instead of user name and primary group name and these identifiers may be different for the same user on different hosts. In that situation these facts helps.

It may be better to change users fact to return Hash of username => uid,gid,homedir,etc instead of Array of usernames though.